README added and template added
